Milestone 9 Update
Added QRDSensor and Solenoid files
Implemented depositBall function
Implemented DEPOSIT_BALL state

TODO: get the depositBall function to be more consistent; may require hardware changes

+++ IRProximitySensor.c
@@ -1,11 +1,11 @@
 #include <xc.h>
 
 typedef struct {
-    volatile unsigned int *readReg;
+    volatile uint16_t *readReg;
     unsigned int bit;
 } IRProximitySensor;
 
-void setupIRProximitySensor(volatile unsigned int *readReg, unsigned int bit, IRProximitySensor *sensor) {
+void setupIRProximitySensor(volatile uint16_t *readReg, unsigned int bit, IRProximitySensor *sensor) {
     sensor->readReg = readReg;
     sensor->bit = bit;
 }
