+++ IRRangeSensor.c
@@ -1,7 +1,9 @@
 #define BUFFER_SIZE 3
 
+#include <stdint.h>
+
 typedef struct { 
-    volatile unsigned int *readBit;
+    volatile uint16_t *readBit;
     volatile unsigned int rangeBuffer[BUFFER_SIZE]; // store a buffer of the last 20 range values
     volatile unsigned int rangeAverage;
     unsigned int lastUpdated;
@@ -27,7 +29,7 @@ unsigned int seesWall(unsigned int withinRangeValue, IRRangeSensor *sensor) {
     return (getRangeValue(sensor) > withinRangeValue);
 }
 
-void setupIRRangeSensor(volatile unsigned int *readBit, IRRangeSensor *sensor) {
+void setupIRRangeSensor(volatile uint16_t *readBit, IRRangeSensor *sensor) {
     sensor->readBit = readBit;
     sensor->rangeAverage = 0;
     sensor->lastUpdated = 0;
