+++ IRRangeSensor.h
@@ -8,10 +8,12 @@
 #ifndef IRRANGESENSOR_H
 #define	IRRANGESENSOR_H
 
+#include <stdint.h>
+
 #define BUFFER_SIZE 3
 
 typedef struct { 
-    volatile unsigned int *readBit;
+    volatile uint16_t *readBit;
     volatile unsigned int rangeBuffer[BUFFER_SIZE]; // store a buffer of the last 20 range values
     volatile unsigned int rangeAverage;
     unsigned int lastUpdated;
@@ -20,7 +22,7 @@ typedef struct {
 void updateRange(IRRangeSensor *sensor);
 unsigned int getRangeValue(IRRangeSensor *sensor);
 unsigned int seesWall(unsigned int withinRangeValue, IRRangeSensor *sensor);
-void setupIRRangeSensor(volatile unsigned int *readBit, IRRangeSensor *sensor);
+void setupIRRangeSensor(volatile uint16_t *readBit, IRRangeSensor *sensor);
 
 #endif	/* IRRANGESENSOR_H */
