+++ StepperMotor.h
@@ -12,12 +12,12 @@ typedef enum {FORWARD, REVERSE} direction;
 
 typedef struct {
     direction direction;
-    volatile unsigned int *dirReg, *pwmPin, *dutyCyclePin;
+    volatile uint16_t *dirReg, *pwmPin, *dutyCyclePin;
     unsigned int countsPerRev, dirBit;
     float rps, wheelDiameter, microstep;
 } StepperMotor;
 
-void setupMotor(direction dir, volatile unsigned int *dirReg, unsigned int dirBit, volatile unsigned int *pwmPin, volatile unsigned int *dutyCyclePin, int countsPerRev, float wheelDiameter, float microstep, StepperMotor *motor);
+void setupMotor(direction dir, volatile uint16_t *dirReg, unsigned int dirBit, volatile uint16_t *pwmPin, volatile uint16_t *dutyCyclePin, int countsPerRev, float wheelDiameter, float microstep, StepperMotor *motor);
 void setDirection(direction dir, StepperMotor *motor);
 void setRPS(float rps, StepperMotor *motor);
 void setSpeed(float speedInPerSec, StepperMotor *motor);
